backend/app/requirements.txt
annotated-types
boto3
anyio
certifi
charset-normalizer
click
fastapi[all]
greenlet
h11
httpcore
httptools
httpx
idna
psycopg2-binary
pydantic
pydantic-settings
pydantic_core
PyJWT
python-dotenv
python-telegram-bot
PyYAML
requests
sniffio
SQLAlchemy
starlette
typing_extensions
urllib3
uvicorn
uvloop
watchfiles
websockets
